tools/devhub-dynamic-dashboards/requirements.txt
fastapi 
uvicorn
pydantic
pyyaml
